requirements.txt
# PHI Redaction Tool Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
presidio-analyzer>=2.2.360
presidio-anonymizer>=2.2.360
spacy>=3.8.0
openpyxl>=3.1.0
pandas>=2.0.0
pyyaml>=6.0

# Language model (install separately)
# python -m spacy download en_core_web_md